move CheckableInput to CheckboxInput components
