+++ src/components/CheckboxInput/components/CheckableInput/CheckableInput.test.tsx
