+++ src/index.ts
@@ -2,9 +2,9 @@ export { default as Accordion } from "./components/Accordion";
 export { default as ActionButton } from "./components/ActionButton";
 export { default as AccordionSection } from "./components/AccordionSection";
 export { default as ArticlePagination } from "./components/ArticlePagination";
+export { default as Badge } from "./components/Badge";
 export { default as Button, ButtonAppearance } from "./components/Button";
 export { default as Card } from "./components/Card";
-export { default as CheckableInput } from "./components/CheckableInput";
 export { default as CheckboxInput } from "./components/CheckboxInput";
 export { default as Chip } from "./components/Chip";
 export { default as Code } from "./components/Code";
@@ -63,9 +63,9 @@ export type {
 } from "./components/AccordionSection";
 export type { ActionButtonProps } from "./components/ActionButton";
 export type { ArticlePaginationProps } from "./components/ArticlePagination";
+export type { BadgeProps } from "./components/Badge";
 export type { ButtonProps } from "./components/Button";
 export type { CardProps } from "./components/Card";
-export type { CheckableInputProps } from "./components/CheckableInput";
 export type { CheckboxInputProps } from "./components/CheckboxInput";
 export type { ChipProps } from "./components/Chip";
 export type { CodeProps } from "./components/Code";
